pyproject.toml
[project]
name = "travel-planning-nemo"
version = "0.1.0"
description = "Travel itinerary generator using NVIDIA NeMo Agent Toolkit (NAT) + NVIDIA API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "nvidia-nat~=1.4",
  "python-dotenv>=1.0.0",
  "pydantic>=2.6.0",
  "tavily-python>=0.5.0",
  "httpx>=0.27.0",
]

[project.entry-points."nat.components"]
travel_planning = "travel_planning_nemo.register"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
